pyproject.toml
[tool.poetry]
name = "darwinsoft-"
version = "0.1.0"
description = "Тестовое задание для DarwinSoft"
authors = ["Kristina Shishkina <klikklok2017@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.31"
fastapi = "^0.111.1"
uvicorn = {extras = ["standart"], version = "^0.30.3"}
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.2.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
